Deprecate `export_table_to_file` to `export_to_file` (#1503)
# Description
closes: https://github.com/astronomer/astro-sdk/issues/1530
Related to:
https://astronomer.slack.com/archives/C03868KGF2Q/p1671621286602989

## What is the current behavior?
Recently, we deprecated `export_file` to `export_table_to_file` but
looks like `export_to_file` is a better name fit for this


## What is the new behavior?
Deprecate  to `export_table_to_file` to  `export_to_file` 

## Does this introduce a breaking change?


### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary
