+++ python-sdk/src/astro/sql/__init__.py
@@ -13,6 +13,7 @@ from astro.sql.operators.dataframe import DataframeOperator, dataframe
 from astro.sql.operators.drop import DropTableOperator, drop_table
 from astro.sql.operators.export_file import ExportFileOperator, export_file
 from astro.sql.operators.export_table_to_file import ExportTableToFileOperator, export_table_to_file
+from astro.sql.operators.export_to_file import ExportToFileOperator, export_to_file
 from astro.sql.operators.load_file import LoadFileOperator, load_file
 from astro.sql.operators.merge import MergeOperator, merge
 from astro.sql.operators.raw_sql import RawSQLOperator, run_raw_sql
@@ -32,6 +33,8 @@ __all__ = [
     "export_file",
     "ExportTableToFileOperator",
     "export_table_to_file",
+    "ExportToFileOperator",
+    "export_to_file",
     "get_value_list",
     "LoadFileOperator",
     "load_file",
