+++ python-sdk/tests/sql/operators/test_export_file.py
@@ -10,8 +10,8 @@ from astro.constants import Database, FileType
 from astro.files import File
 
 # Import Operator
-from astro.sql import ExportFileOperator, export_file
-from astro.sql.operators.export_table_to_file import ExportTableToFileOperator, export_table_to_file
+from astro.sql import ExportFileOperator, ExportTableToFileOperator, export_file, export_table_to_file
+from astro.sql.operators.export_to_file import ExportToFileOperator, export_to_file
 from astro.table import Table
 
 from ..operators import utils as test_utils
@@ -26,7 +26,7 @@ def test_save_dataframe_to_local(sample_dag):
 
     with sample_dag:
         df = make_df()
-        aql.export_table_to_file(
+        aql.export_to_file(
             input_data=df,
             output_file=File(path="/tmp/saved_df.csv"),
             if_exists="replace",
@@ -66,7 +66,7 @@ def test_save_returns_output_file(sample_dag, database_table_fixture):
     data_path = str(CWD) + "/../../data/homes.csv"
     with sample_dag:
         table = aql.load_file(input_file=File(path=data_path), output_table=test_table)
-        file = aql.export_table_to_file(
+        file = aql.export_to_file(
             input_data=table,
             output_file=File(path="/tmp/saved_df.csv"),
             if_exists="replace",
@@ -109,14 +109,14 @@ def test_unique_task_id_for_same_path(
 
             if i == 3:
                 params["task_id"] = "task_id"
-            task = export_table_to_file(**params)
+            task = export_to_file(**params)
             tasks.append(task)
     test_utils.run_dag(sample_dag)
 
     assert tasks[0].operator.task_id != tasks[1].operator.task_id
-    assert tasks[0].operator.task_id == "export_table_to_file"
-    assert tasks[1].operator.task_id == "export_table_to_file__1"
-    assert tasks[2].operator.task_id == "export_table_to_file__2"
+    assert tasks[0].operator.task_id == "export_to_file"
+    assert tasks[1].operator.task_id == "export_to_file__1"
+    assert tasks[2].operator.task_id == "export_to_file__2"
     assert tasks[3].operator.task_id == "task_id"
 
     os.remove(OUTPUT_FILE_PATH)
@@ -127,7 +127,7 @@ def test_inlets_outlets_supported_ds():
     """Test Datasets are set as inlets and outlets"""
     input_data = Table("test_name")
     output_file = File("gs://bucket/object.csv")
-    task = aql.export_table_to_file(
+    task = aql.export_to_file(
         input_data=input_data,
         output_file=output_file,
     )
@@ -140,7 +140,7 @@ def test_inlets_outlets_non_supported_ds():
     """Test inlets and outlets are not set if Datasets are not supported"""
     input_data = Table("test_name")
     output_file = File("gs://bucket/object.csv")
-    task = aql.export_table_to_file(
+    task = aql.export_to_file(
         input_data=input_data,
         output_file=output_file,
     )
@@ -149,9 +149,9 @@ def test_inlets_outlets_non_supported_ds():
 
 
 def test_raise_exception_for_invalid_input_type():
-    """Raise value error when input data is not correct in ExportTableToFileOperator"""
+    """Raise value error when input data is not correct in ExportToFileOperator"""
     with pytest.raises(ValueError) as exc_info:
-        ExportTableToFileOperator(
+        ExportToFileOperator(
             task_id="task_id",
             input_data=123,
             output_file=File(
@@ -166,12 +166,12 @@ def test_raise_exception_for_invalid_input_type():
 
 
 # TODO: Remove this test in astro-sdk 2.0.0
-def test_warnings_message():
+def test_export_file_warnings_message():
     """Assert the warning log when using deprecated method"""
     with pytest.warns(
         expected_warning=DeprecationWarning,
         match="""This class is deprecated.
-            Please use `astro.sql.operators.export_table_to_file.ExportTableToFileOperator`.
+            Please use `astro.sql.operators.export_to_file.ExportToFileOperator`.
             And, will be removed in astro-sdk-python>=2.0.0.""",
     ):
         ExportFileOperator(
@@ -188,7 +188,38 @@ def test_warnings_message():
     with pytest.warns(
         expected_warning=DeprecationWarning,
         match="""This decorator is deprecated.
-        Please use `astro.sql.operators.export_table_to_file.export_table_to_file`.
+        Please use `astro.sql.operators.export_to_file.export_to_file`.
         And, will be removed in astro-sdk-python>=2.0.0.""",
     ):
         export_file(input_data=Table(), output_file=File(path="/tmp/saved_df.csv"), if_exists="replace")
+
+
+# TODO: Remove this test in astro-sdk 1.5.0
+def test_export_table_to_file_warnings_message():
+    """Assert the warning log when using deprecated method"""
+    with pytest.warns(
+        expected_warning=DeprecationWarning,
+        match="""This class is deprecated.
+            Please use `astro.sql.operators.export_to_file.ExportToFileOperator`.
+            And, will be removed in astro-sdk-python>=1.5.0.""",
+    ):
+        ExportTableToFileOperator(
+            task_id="task_id",
+            input_data=123,
+            output_file=File(
+                path="gs://astro-sdk/workspace/openlineage_export_file.csv",
+                conn_id="bigquery",
+                filetype=FileType.CSV,
+            ),
+            if_exists="replace",
+        )
+
+    with pytest.warns(
+        expected_warning=DeprecationWarning,
+        match="""This decorator is deprecated.
+        Please use `astro.sql.operators.export_to_file.export_to_file`.
+        And, will be removed in astro-sdk-python>=1.5.0.""",
+    ):
+        export_table_to_file(
+            input_data=Table(), output_file=File(path="/tmp/saved_df.csv"), if_exists="replace"
+        )
